chore: create a TimeDisplay utils and put reformatTime in it
